requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
Django==6.0.4
django-oauth-toolkit==3.2.0
djangorestframework==3.17.1
idna==3.13
jwcrypto==1.5.7
oauthlib==3.3.1
psycopg2-binary==2.9.12
pycparser==3.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
